pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "sofia-utils"
version = "1.0.0"
description = "A utilities package named after my beautiful siamese cat"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [
  { name = "Luis Ignacio Reyes Castro", email = "luis.i.reyes.castro@gmail.com" },
]
dependencies = [
  "Babel>=2.17.0",
  "GitPython>=3.1.45",
  "phonenumbers>=9.0.13",
  "tiktoken>=0.9.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sofia_utils*"]
